Remove duplicated header and use a common header
