+++ src/pages/home/HomePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import MainTemplate from '../../components/main/MainTemplate';
-import MainHeader from '../../components/main/MainHeader';
+import Header from '../../components/base/Header';
 import HomeTab from '../../components/home/HomeTab';
 import MainResponsive from '../../components/main/MainResponsive';
 import HomeLayout from '../../components/home/HomeLayout';
@@ -8,15 +8,15 @@ import { Route } from 'react-router-dom';
 import TrendingPostsPage from './TrendingPostsPage';
 import RecentPostsPage from './RecentPostsPage';
 import HomeSidebar from '../../components/home/HomeSidebar';
-import FloatingMainHeader from '../../components/main/FloatingHomeHeader';
+import FloatingHeader from '../../components/base/FloatingHeader';
 
 export type HomePageProps = {};
 
 function HomePage(props: HomePageProps) {
   return (
     <MainTemplate>
-      <MainHeader />
-      <FloatingMainHeader />
+      <Header />
+      <FloatingHeader />
       <MainResponsive>
         <HomeTab />
         <HomeLayout
